feat: :sparkles: Adiciona a entidade pedido e a rota post no seu controller

+++ src/main/java/br/com/lanchebom/models/repository/AdicionalRepository.java
@@ -3,5 +3,5 @@ package br.com.lanchebom.models.repository;
 import br.com.lanchebom.models.entity.Adicional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdicionalRepository extends JpaRepository<Adicional, Integer> {
+public interface AdicionalRepository extends JpaRepository<Adicional, Long> {
 }
