+++ src/main/java/br/com/lanchebom/models/repository/LancheRepository.java
@@ -3,5 +3,5 @@ package br.com.lanchebom.models.repository;
 import br.com.lanchebom.models.entity.Lanche;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LancheRepository extends JpaRepository<Lanche, Integer> {
+public interface LancheRepository extends JpaRepository<Lanche, Long> {
 }
